+++ wemake_python_styleguide/presets/topics/complexity.py
@@ -28,6 +28,7 @@ PRESET: Final = (
     counts.TryExceptVisitor,
     counts.ReturnLikeStatementTupleVisitor,
     counts.TupleUnpackVisitor,
+    counts.TypeParamsVisitor,
     classes.ClassComplexityVisitor,
     classes.MethodMembersVisitor,
     overuses.StringOveruseVisitor,